Refactor building logic out of RepoHierarchy into a builder
